tests are moved to avoid package namespace pollution
