+++ lib/go/src/thrift/tests/gen-go/librarytest/EmptyService/EmptyService-remote.go
