+++ lib/go/src/thrift/tests/gen-go/librarytest/ServiceForExceptionWithAMap/ServiceForExceptionWithAMap-remote.go
