+++ lib/go/src/thrift/tests/gen-go/librarytest/Srv/Srv-remote.go
