+++ lib/go/src/thrift/tests/gen-go/lists/ttypes.go
@@ -6,8 +6,8 @@ package lists
 
 import (
 	"fmt"
+	"github.com/apesternikov/thrift4go/lib/go/src/thrift"
 	"math"
-	"thrift"
 )
 
 // This is a temporary safety measure to ensure that the `math'
