+++ lib/go/src/thrift/tests/maps.thrift
