+++ lib/go/src/thrift/tests/tbinary_protocol_test.go
@@ -20,22 +20,10 @@
 package thrift
 
 import (
+	. "github.com/apesternikov/thrift4go/lib/go/src/thrift"
 	"testing"
 )
 
-func TestTApplicationException(t *testing.T) {
-	exc := NewTApplicationException(UNKNOWN_APPLICATION_EXCEPTION, "")
-	if exc.Error() != "" {
-		t.Fatalf("Expected empty string for exception but found '%s'", exc.Error())
-	}
-	if exc.TypeId() != UNKNOWN_APPLICATION_EXCEPTION {
-		t.Fatalf("Expected type UNKNOWN for exception but found '%s'", exc.TypeId())
-	}
-	exc = NewTApplicationException(WRONG_METHOD_NAME, "junk_method")
-	if exc.Error() != "junk_method" {
-		t.Fatalf("Expected 'junk_method' for exception but found '%s'", exc.Error())
-	}
-	if exc.TypeId() != WRONG_METHOD_NAME {
-		t.Fatalf("Expected type WRONG_METHOD_NAME for exception but found '%s'", exc.TypeId())
-	}
+func TestReadWriteBinaryProtocol(t *testing.T) {
+	ReadWriteProtocolTest(t, NewTBinaryProtocolFactoryDefault())
 }
