+++ lib/go/src/thrift/tests/tframed_transport_test.go
@@ -20,17 +20,11 @@
 package thrift
 
 import (
-	"io"
+	. "github.com/apesternikov/thrift4go/lib/go/src/thrift"
 	"testing"
 )
 
-func TestTException(t *testing.T) {
-	exc := NewTException("")
-	if exc.Error() != "" {
-		t.Fatalf("Expected empty string for exception but found '%s'", exc.Error())
-	}
-	exc = NewTExceptionFromOsError(io.EOF)
-	if exc.Error() != io.EOF.Error() {
-		t.Fatalf("Expected '%s', but found '%s'", io.EOF.Error(), exc.Error())
-	}
+func TestFramedTransport(t *testing.T) {
+	trans := NewTFramedTransport(NewTMemoryBuffer())
+	TransportTest(t, trans, trans)
 }
