+++ lib/go/src/thrift/tests/tnonblocking_transport_test.go
@@ -20,6 +20,7 @@
 package thrift
 
 import (
+	. "github.com/apesternikov/thrift4go/lib/go/src/thrift"
 	"net"
 	"testing"
 )
